[FIX] optimize new rule
