--- account_advanced_reconcile_bank_statement/base_advanced_reconciliation.py
@@ -1,46 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Matthieu Dietrich
-#    Copyright 2014 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from itertools import product
-from openerp.osv import orm
-
-
-class easy_reconcile_advanced(orm.AbstractModel):
-
-    _inherit = 'easy.reconcile.advanced'
-
-    def _base_columns(self, rec):
-        """ Mandatory columns for move lines queries
-        An extra column aliased as ``key`` should be defined
-        in each query."""
-        aml_cols = (
-            'id',
-            'debit',
-            'credit',
-            'date',
-            'period_id',
-            'ref',
-            'name',
-            'partner_id',
-            'account_id',
-            'move_id',
-            'statement_id')
-        return ["account_move_line.%s" % col for col in aml_cols]
